Use the k-d tree in ComponentComsol to match up potentials and mesh nodes.

+++ Include/Garfield/KDTree.hh
@@ -25,7 +25,7 @@ class KDTreeNode;
 
 struct KDTreeResult {
   double dis; //< square Euclidean distance
-  int idx;    //< index
+  size_t idx; //< index
 }; 
 
 /// Main k-d tree class.
